Core SDKs (Node/Python/Go/.NET) + generated types, codegen, docs, harness, and workflows updated to upstream. Kept our structure for java/rust and the 34 additional SDKs; kept our supercharged README/CHANGELOG/copilot-instructions. New features to port: rewind, agent factories authoring, in-process FFI transport, reasoningEffort:max, plugin dirs, additional session directories, decisionContext, disabled MCP servers, githubMcpToolConfig, userPromptTransformed hook, and more.

+++ dotnet/test/E2E/RpcWorkspaceCheckpointsE2ETests.cs
@@ -29,7 +29,7 @@ public class RpcWorkspaceCheckpointsE2ETests(E2ETestFixture fixture, ITestOutput
     {
         await using var session = await CreateSessionAsync();
 
-        var result = await session.Rpc.Workspaces.ReadCheckpointAsync(long.MaxValue);
+        var result = await session.Rpc.Workspaces.ReadCheckpointAsync(uint.MaxValue);
 
         Assert.True(string.IsNullOrEmpty(result.Content));
     }
